Save work

+++ src/control/deleteCollection.ts
@@ -1,7 +1,8 @@
 import { ManageIndexesApi } from '../pinecone-generated-ts-fetch/control';
 import { buildConfigValidator } from '../validator';
-import { CollectionNameSchema } from './types';
+// import { CollectionNameSchema } from './types';
 import type { CollectionName } from './types';
+import { PineconeArgumentError } from '../errors';
 
 /**
  * The name of collection to delete.
@@ -9,14 +10,13 @@ import type { CollectionName } from './types';
 export type DeleteCollectionOptions = CollectionName;
 
 export const deleteCollection = (api: ManageIndexesApi) => {
-  const validator = buildConfigValidator(
-    CollectionNameSchema,
-    'deleteCollection'
-  );
-
   return async (collectionName: DeleteCollectionOptions): Promise<void> => {
-    validator(collectionName);
-
+    if (!collectionName || collectionName.length === 0) {
+      throw new PineconeArgumentError(
+        'You must pass a non-empty string for `collectionName` in order to delete a' +
+          ' collection'
+      );
+    }
     await api.deleteCollection({ collectionName });
     return;
   };
